+++ packages/terra-form-textarea/tests/nightwatch/textarea-spec.js
@@ -2,7 +2,7 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 const resizeTo = require('terra-toolkit/lib/nightwatch/responsive-helpers').resizeTo;
 
-module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], {
+module.exports = resizeTo(['tiny', 'huge'], {
   'Displays a default Textarea': (browser) => {
     browser.url(`${browser.launchUrl}/#/tests/form-textarea-tests/default`);
     browser.expect.element('textarea').to.be.present;