+++ packages/terra-image/tests/nightwatch/image-spec.js
@@ -1,7 +1,7 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 const resizeTo = require('terra-toolkit/lib/nightwatch/responsive-helpers').resizeTo;
 
-module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], {
+module.exports = resizeTo(['tiny', 'huge'], {
   'Displays an image with default options': (browser) => {
     browser.url(`${browser.launchUrl}/#/tests/image-tests/default`);
   },