+++ packages/terra-menu/tests/nightwatch/menu-item-group/menu-item-group-spec.js
@@ -1,7 +1,7 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 const resizeTo = require('terra-toolkit/lib/nightwatch/responsive-helpers').resizeTo;
 
-module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], {
+module.exports = resizeTo(['tiny', 'huge'], {
   'Displays a default item group': (browser) => {
     browser
       .url(`${browser.launchUrl}/#/tests/menu-item-group-tests/default`)