+++ packages/terra-overlay/tests/nightwatch/Overlay/overlay-spec.js
@@ -2,7 +2,7 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 const resizeTo = require('terra-toolkit/lib/nightwatch/responsive-helpers').resizeTo;
 
-module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], {
+module.exports = resizeTo(['tiny', 'huge'], {
   'Displays a default overlay': (browser) => {
     browser.url(`${browser.launchUrl}/#/tests/overlay-tests/default`);