+++ packages/terra-site/src/examples/form-radio/DefaultRadioExample.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import Radio from 'terra-form-radio/lib/Radio';
+
+const defaultRadioExample = () => (<Radio id="default-radio" labelText="Default Radio" name="default" />);
+
+export default defaultRadioExample;