+++ packages/terra-site/src/examples/form-radio/HiddenLabelRadioExample.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import Radio from 'terra-form-radio/lib/Radio';
+
+const hiddenLabelRadioExample = () => (<Radio id="hidden-radio" labelText="can you see me?" isLabelHidden />);
+
+export default hiddenLabelRadioExample;