+++ packages/terra-site/src/examples/image/ImageDefault.jsx
@@ -19,6 +19,12 @@ const ImageDefault = () => (
     <br />
     <h2>Fluid</h2>
     <Image src={placeholderPic1000x200} alt="fluid image" isFluid />
+    <br />
+    <h2>Placeholder</h2>
+    <h3>Invalid src prop (placeholder will show)</h3>
+    <Image src={'/mock/failure/route'} placeholderSrc={placeholderPic1000x200} alt="placeholder image" />
+    <h3>Valid src prop (src will show)</h3>
+    <Image src={placeholderPic150x150} placeholderSrc={placeholderPic1000x200} alt="src image" />
   </div>
 );