+++ packages/terra-hookshot/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+1.7.0 - (January 18, 2018)
+------------------
+### Changed
+* Minor version bump
+
 1.6.0 - (January 5, 2018)
 ------------------
 ### Changed