+++ packages/terra-list/CHANGELOG.md
@@ -3,8 +3,12 @@ Changelog
 
 Unreleased
 ----------
+
+1.19.0 - (January 18, 2018)
+------------------
 ### Changed
 * Add context to selectable list nightwatch example
+* Move selectable functions to utils and convert to selectable list to functional component
 
 1.18.0 - (January 5, 2018)
 ------------------