+++ packages/terra-tabs/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 -----------------
 
+1.1.0 - (January 18, 2018)
+------------------
+### Changed
+* Minor version bump
+
 1.0.0 - (January 5, 2018)
 ------------------
 * Initial stable release